Avoids situations where there is no changes

+++ src/HodStudio.EntityFrameworkDiffLog.TestsDotNet45/Model/Department.cs
@@ -8,6 +8,6 @@ namespace HodStudio.EntityFrameworkDiffLog.TestsDotNet45.Model
     {
         public Guid Id { get; set; } = Guid.NewGuid();
         public Guid Code { get; set; }
-        public decimal Budget { get; set; }
+        public long Budget { get; set; }
     }
 }
